Fixes the consequences when saving empty file data.
When the file field is allowed to be empty and the storage entity is saved without a file present it has still saved successfully even in the case no file data was present. It also created an empty file in the storage backend (using the local adapter when I discovered this).

This commit should resolve these issues and improve the storage event system a little more.