Improving the storage event listeners.
There is now an exception throw if a listener can't work with a specific adapter. Also I've introduced a new BaseListener that doesn't have any restrictions on which adapter it will take. In theory you can use it with any storage backend as long as the path it builds for the files is compatible with your chosen storage backend.

A new shell command has been introduced to store a file via command line.

The Quick Start tutorial has been updated as well.